requirements.txt
#requirements.txt

# Core dependencies
python>=3.8,<3.12
streamlit>=1.28.0
plotly>=5.15.0
pandas>=2.0.0
numpy>=1.21.0
xarray>=2023.1.0
netCDF4>=1.6.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
python-dotenv>=1.0.0

# NLP and AI
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0
scikit-learn>=1.2.0
nltk>=3.8.0

# Visualization
folium>=0.14.0
geopandas>=0.13.0
shapely>=2.0.0
pyproj>=3.5.0

# Data processing
scipy>=1.10.0
dask>=2023.1.0
zarr>=2.14.0

# Web and API
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
requests>=2.28.0

# Utilities
pyyaml>=6.0
click>=8.1.0
tqdm>=4.65.0
colorama>=0.4.0
python-dateutil>=2.8.0

# Development and testing
pytest>=7.3.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Optional (performance)
psutil>=5.9.0
memory-profiler>=0.60.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.0.0


matplotlib
folium
streamlit-folium
netcdf4
pillow
requests